Refactor entrypoints, add requirements.txt and update README
