+++ README.md
@@ -1,10 +1,17 @@
 # entree-plat-dessert
 Stable marriages implementation done for the HAI904I course of the University of Montpellier by 4 students (master's degree).
 
-# Interface
-We use Tkinter to create a graphic interface for our tool. More specifically, we use the [CustomTkinter](https://github.com/TomSchimansky/CustomTkinter) python library.
-To make this interface work on any desktop platform you first need to install **CustomTkinter** and **packaging** in case you don't have it.
+# Installation and use
+
+This project requires Python 3. Python modules can be installed with pip :
+
 ```
-pip3 install customtkinter
-pip3 install packaging
+git clone git@github.com:dym-dym/entree-plat-dessert.git
+cd entree-plat-dessert
+pip install -r requirements.txt
+```
+
+To run the main application interface :
+```
+python main.py
 ```
